requirements.txt
Flask
Gunicorn
flask_socketio
openai
logging
haystack.telemetry
typing
json
pprint
